requirements.txt
python-telegram-bot==20.7
pyrogram==2.0.106
pymongo==4.6.0
motor==3.3.2
APScheduler==3.10.4
redis==5.0.1
aiohttp==3.9.1
python-magic==0.4.27
Pillow==10.1.0
python-dotenv==1.0.0
tgcrypto==1.2.5
uvloop==0.19.0
